Move to separate files & write docs
